+++ Graveyard Keeper/src/GerrysJunkTrunk/CHANGELOG.md
@@ -1,8 +1,9 @@
 # Changelog
 
-## 1.9.7 | 16 April 2026
+## 1.9.7 | 17 April 2026
 
 - Fixed Chinese translations not loading
+- Added an update notice on the main menu that flags when this mod (or others in this collection) has a newer version on Nexus. Click an entry to open its Nexus page. Toggle off in settings if you'd rather not see it
 
 ## 1.9.6 | 15 April 2026